@@ -9,12 +9,12 @@
     "@testing-library/user-event": "^12.8.3",
     "dom-event-target": "^1.0.0",
     "node-sass": "^6.0.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-clipboard.js": "^2.0.16",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-router": "^6.0.0",
     "react-router-dom": "^6.0.0",
-    "react-scripts": "5.0.0",
+    "react-scripts": "5.0.1",
     "tone": "^14.7.77",
     "web-vitals": "^1.1.2"
   },
@@ -47,8 +47,8 @@
   "devDependencies": {
     "@types/jest": "^27.0.1",
     "@types/node": "^16.9.4",
-    "@types/react": "^17.0.21",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
     "@types/webmidi": "^2.0.6",
     "prettier": "^2.5.1",
     "teaful": "^0.9.2",
